Refactor dashboard layout: simplify structure, remove unused components, and enhance styling
