+++ apps/nextjs/src/app/(dashboard)/dashboard/(body)/tournaments/page.tsx
@@ -2,10 +2,12 @@ import ComingSoon from "~/components/coming-soon";
 
 export default function TournamentHistory() {
   return (
-    <ComingSoon title="Tournament History">
-      <div className="flex flex-col items-center">
-        <h2>Tournament History is under construction</h2>
-      </div>
-    </ComingSoon>
+    <div className="m-4 flex flex-col px-80 py-8">
+      <ComingSoon title="Tournament History">
+        <div className="flex flex-col items-center">
+          <h2>Tournament History is Under Construction</h2>
+        </div>
+      </ComingSoon>
+    </div>
   );
 }
