--- apps/nextjs/src/app/(dashboard)/dashboard/_components/content/dashboard-content-body.tsx
@@ -1,9 +1,0 @@
-import type { ChildrenProps } from "~/types";
-
-export function DashboardContentBody({ children }: ChildrenProps) {
-  return (
-    <div className="min-h-[100dvh] flex-1 rounded-xl bg-neutral-900/50 md:min-h-min">
-      {children}
-    </div>
-  );
-}
