+++ apps/nextjs/src/app/(dashboard)/dashboard/_components/content/dashboard-content-skeleton.tsx
@@ -1,4 +1,4 @@
-import { DashboardContentSection } from "./dashboard-content-top-section";
+import { DashboardContentSection } from "./dashboard-content-section";
 
 export default function DashboardContentSkeleton() {
   return (
