+++ apps/nextjs/src/components/coming-soon.tsx
@@ -9,14 +9,10 @@ export default function ComingSoon({
   children,
 }: Readonly<ComingSoonProps>) {
   return (
-    <div className="flex min-h-screen flex-col pt-4">
-      <div className="animate-fade-in space-y-6 rounded-xl bg-white/10 p-8 text-center shadow-xl backdrop-blur-lg">
-        <div className="space-y-2">
-          <h1 className="text-4xl font-bold">{title}</h1>
-          {children}
-        </div>
-        <p className="text-xl">Coming Soon</p>
-      </div>
+    <div className="animate-fade-in space-y-6 rounded-xl bg-white/10 p-8 text-center shadow-xl backdrop-blur-lg">
+      <h1 className="text-4xl font-bold">{title}</h1>
+      {children}
+      <p className="text-xl">Coming Soon</p>
     </div>
   );
 }
